+++ src/merlin.ml
@@ -11,12 +11,15 @@ type t =
   ; libname    : string option
   }
 
-let ppx_flags sctx ~dir ~src_dir { preprocess; libname; _ } =
+(* This must be forced after we change the cwd to the workspace root *)
+let root_absolute_name = lazy (Sys.getcwd ())
+
+let ppx_flags sctx ~dir ~src_dir:_ { preprocess; libname; _ } =
   match preprocess with
   | Pps { pps; flags } ->
     let exe = SC.PP.get_ppx_driver sctx pps ~dir ~dep_kind:Optional in
     let command =
-      List.map (Path.reach exe ~from:src_dir
+      List.map (Filename.concat (Lazy.force root_absolute_name) (Path.to_string exe)
                 :: "--as-ppx"
                 :: SC.PP.cookie_library_name libname
                 @ flags)